writing-comp/requirements.txt
streamlit==1.31.1
boto3==1.34.29
python-dotenv==1.0.0
streamlit-drawable-canvas==0.9.3
Pillow==10.2.0
transformers==4.36.2
torch==2.2.0
numpy==1.26.4
